requirements.txt
azure-ai-inference[opentelemetry]==1.0.0b9
opentelemetry-sdk==1.34.1
opentelemetry-api
opentelemetry-exporter-otlp-proto-http==1.34.1
python-dotenv